requirements.txt
pillow
mongoengine
flask
pyyaml
python-consul
dnspython